refactor: menu and page component
