+++ packages/frontend/src/pages/notifications.tsx
@@ -1,75 +1,95 @@
-import { RetryError } from '@/components/global/RetryError';
 import Spinner from '@/components/global/Spinner';
 import useError from '@/hooks/useError';
 import useNotifications from '@/hooks/useNotifications';
+import { UserContextType } from '@/types/components';
+import { useContext, useMemo, useState } from 'react';
+import { UserContext } from './_app';
+import { SingleNotification } from '@/components/notifications/SingleNotification';
+import { MarkAllAsRead, setNotificationAsRead } from '@/components/notifications/Notifications';
+import { useAccount } from 'wagmi';
+import { Filters } from '@/components/post/Filters';
+import { faCheck } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faReply, faReplyAll, faCircleUp } from '@fortawesome/free-solid-svg-icons';
-import { NotificationType } from '@/types/components';
-
-const getNotificationIconFromType = (type: NotificationType) => {
-  switch (type) {
-    case NotificationType.DirectReply:
-      return <FontAwesomeIcon icon={faReplyAll} />;
-    case NotificationType.DiscussionReply:
-      return <FontAwesomeIcon icon={faReply} />;
-    default:
-      return <FontAwesomeIcon icon={faCircleUp} color={'#0e76fd'} />;
-  }
-};
+import { RefreshNotifications } from '@/components/notifications/RefreshNotifications';
 
 export default function Notifications() {
+  const { address } = useAccount();
+  const { pushRoute, nymOptions } = useContext(UserContext) as UserContextType;
   const { errorMsg, setError } = useError();
+  const { notifications, unreadNotifications, setNotifications, isLoading } = useNotifications({
+    enabled: true,
+  });
 
-  const { notifications, isLoading } = useNotifications({ enabled: true });
+  const [filter, setFilter] = useState('all');
+
+  const notificationsToShow = useMemo(
+    () => (filter === 'unread' ? unreadNotifications : notifications),
+    [filter, notifications, unreadNotifications],
+  );
+
+  const filters = {
+    all: 'All',
+    unread: 'Unread',
+  };
 
   return (
     <main>
       <div className="w-full bg-gray-50 flex flex-col justify-center items-center">
         <div className="bg-gray-50 min-h-screen w-full">
           <div className="flex flex-col gap-8 max-w-3xl mx-auto py-5 md:py-10 px-3 md:px-0">
-            <div className="flex justify-between">
+            <div className="flex justify-between items-center">
               <h3>Notifications</h3>
+              <RefreshNotifications nymOptions={nymOptions} />
             </div>
             {isLoading ? (
               <Spinner />
-            ) : notifications ? (
+            ) : notificationsToShow ? (
               <>
-                {notifications && notifications.length > 0 ? (
-                  notifications.map((n, i) => {
-                    return n.type === NotificationType.Upvote ? (
-                      <div className="flex gap-4 items-center" key={i}>
-                        {getNotificationIconFromType(n.type)}
-                        <a href={`/users/${n.userId}`}>{n.userId}</a>
-                        <a href={`/posts/${n.id}`}>{n.body}</a>
-                      </div>
-                    ) : (
-                      <div className="flex gap-4 items-center" key={i}>
-                        {getNotificationIconFromType(n.type)}
-                        <a href={`/users/${n.userId}`}>{n.userId}</a>
-                        <div className="flex-col gap-2">
-                          <p>
-                            <a href={`/posts/${n.id}`}>{n.body}</a>
-                          </p>
-                          <p>
-                            <sub>
-                              <a href={`/posts/${n.id}`}>{n.title}</a>
-                            </sub>
-                          </p>
-                        </div>
-                      </div>
-                    );
-                  })
+                <div className="flex justify-between items-center">
+                  <Filters
+                    filters={filters}
+                    selectedFilter={filter}
+                    setSelectedFilter={setFilter}
+                  />
+                  <div
+                    className="flex gap-1 justify-end items-center"
+                    onClick={() =>
+                      MarkAllAsRead(address as string, notifications, setNotifications)
+                    }
+                  >
+                    <FontAwesomeIcon icon={faCheck} size={'xs'} />
+                    <p className="secondary hover:underline">Mark all as read</p>
+                  </div>
+                </div>
+                {notificationsToShow && notificationsToShow.length > 0 ? (
+                  notificationsToShow.map((n, i) => (
+                    <div
+                      className="flex gap-4 justify-between outline-none rounded-2xl transition-all shadow-sm bg-white p-3 md:px-5 md:py-4 border border-gray-200 hover:border-gray-300 hover:cursor-pointer w-full"
+                      key={i}
+                      onClick={() => {
+                        setNotificationAsRead(
+                          address as string,
+                          notifications,
+                          setNotifications,
+                          n.id,
+                        );
+                        pushRoute(`/posts/${n.id}`);
+                      }}
+                    >
+                      <SingleNotification n={n} />
+                    </div>
+                  ))
+                ) : isLoading ? (
+                  <div className="p-4">
+                    <Spinner />
+                  </div>
                 ) : (
-                  <p className="text-center">No notifications</p>
+                  <div className="m-auto">
+                    <p>No notifications.</p>
+                  </div>
                 )}
               </>
-            ) : (
-              <RetryError
-                message="Could not fetch users."
-                error={errorMsg}
-                refetchHandler={() => {}}
-              />
-            )}
+            ) : null}
           </div>
         </div>
       </div>
